Fix endless loops, useless retries and skipped back-off increment during API calls
GraphQL API calls had several problems, as they repeated calls that contained implementation issues, never stopped retrying api calls (leading to leaks) and in case of exception skipped the incrementation of back-off time.
This commit attempts to fix all issues and refactored the error handling a bit, aiming it collect it to fewer places. Ideally swallowing the least amount of errors in the lower parts of the call stack.

refs #12818