Normalize invocation of ApiService functions.
Reduce duplication of logic by removing the "Async" version of the api call functions.
Async was misleading to, as non-async functions also where async. What actually happened is that Async jobs
detached it in a job polling indefinetly. To make this behavior more transparent you now have to wrap your function in a retryApiCall block.