Fix inconsistent state of bookmark list
Due to multiple state holders of current bookmarks in recycler view we might end up in an inconsistent state. bookmarked articles are written by live data observer and peculiarly from a viewholder conserving that state.
A workaround seems to always write the bookmarks before delete, but I added a todo to clean this up.

refs #12935