Assortment of performance improvements
*Encountered deadlocks while accessing downloadLiveData in DataService*

`synchronized` actually blocks threads instead of coroutines, so in massive concurrency situations we might end up blocking all coroutine dispatcher threads

*Replace okHttp with ktor http client*

The reasoning behind this it supports a coroutine-io based connection backend, meaning we can gracefully cancel requests via coroutine api. Benefits performance in RecyclerView where MomentViewDataBinding attempts to download moments as the user scrolls.
If the user scrolls past that download becomes obsolete and can be cancelled.

*Hide web view if moment is not animated*

To save the rendering time of the web view in MomentView we hide it if it's not needed

*Scrap all lock based download limitation strategies*

We don't need to lock our downloading mechanisms to a specific limit manually as the HttpClient supports connection limits out of the box. We can leave it up to it.

*Rework file integrity management*

Checking if files exist is extremely error prone, for each collection we analyze which files are not meeting integrity (either through not being downloaded yet or be somehow corrupted by deviating from the expected sha sum).
We only attempt to re-download the corrupt files and fail softly.

*Make Moments standalone*

A major perfomance hit was taken due to the fact we immediately download the whole metadata of an issue (with all its joins) if we scrolled by although we only needed the Moment and its files. So I decoupled the Moment from the Issue and introduced
methods to download and save Moments individually without having to download an Issue altogether. Issue download is only triggered now if opening it in the IssueContentFragment.

*Misc*

* bump mockito
* Enhance test coverage for IssueDownloadWorkManagerWorker
* Enhance test coverage for NewIssuePollingScheduler