Eliminate unsafe singleton access wherever context available
As we have sporadic crashes reported in sentry where a Singleton might not be initialized in certain situations it seems to be a good idea to
pass a context wherever possible. There are classes from where a context is not easily accessible where i would propose refactor to a proper DI solution.
However this should fix 99% of all singleton-lifecycle related issues.