Avoid implicit storage path determination by introducing a user-selectable storage option
In the past the storage was determined by the availability of "getExternalFilesDir()", which was used automatically.
This also could change during runtime if i.e. user mounts a SD-Card. Therefore this MR introduces an explicitly configurable storage.

FileEntries now carry its storage location (EXTERNAL, INTERNAL, NOT_STORED) and its relative path.
If there are downloaded items that are not stored on the configured storage device the migration activity is triggered at startup copying all files to the new, correct location.