Making the getIssue and getMoment functions status agnostic
While delivering Issues with status "regular" to non-authenticated users I realized it's a bad idea to predetermine the
expected status of the issue by the users authstatus.

Instead all getIssue / getMoment functions will do a hierachichal lookup preferring the regular version - if not found try for the public in cache.
After that the funciton will hit the server anyway returning whatever issue is available. I realized we need a lot of yanky logic as this violates stateless transfer principles... welp.

Also I found and fixed another but - if you have files that are downloaded by a public issue and are referenced on a regular as well it's placed in the "wrong" folder. so some assets remain in the public folder while they are needed
by a regular issue. That was solvable though by just determining the actual path while overriding urls in the webview, so it doesn't matter anymore in which folder the assets are.

refs #13095