So i figured something out:
The primary external Storage could be a phone partition that is "external" or, on device that don't do that the said SD-Card. If you have "internal" external storage the primary external storage will never be an SD-Card. Probably that's the source of confusion on FP2.
I think we should continue only the primary external storage device whether that is your "internal" external storage or a removable SD-Card.
Existing users won't be affected by this change, as we continue to only use _primary_ external storage. Users that have "internal" external storage will gain the ability to use that.
The help text tries to communicate this to our users.

I hope now everything surrounding this annoying topic is cleared up.