Fix more inconsistencies while determining the download status of issues with or without pages
Furthermore deleted the mechanism that download live datas of issues with pages might be taken for a live data for a issue without. That was originally intended for displaying the download status on a regular issue if a pdf one already is downloading. That lead to a lot of confusion and weird state in the download observer. Easiest fix is to delete that.
The bug is fixed while we only "lose" the download status indicator being started on a regular issue while the pdf one is downloading