Add query definition, data model and little poc for search endpoints
This commit adds the interfaces and a function in ApiService to query the search in the
graphql interface. Basic data types are included and a experimental activity
as a little poc was introduced.

I thought it's cool to have a somewhat hidden section in the settings where we can
expose debugging options / activities (debug build only)