Refactor cache operations (Metadata download, Content download)
So here it is - my attempt to improve readability, performance and reliability of our downloading mechanism
I tried to compensate the monstrous commit with lots of docstrings and comments.

Some prose to get yourself an overview:

We have a big new class the "CacheOperation". It serves as an observable process of individual, volatile, slow IO bound items being processed in the name of a tag (typically a "getDownloadTag" from some collection).
In its companion object there is a synchronized list keeping track of any ongoing cacheoperation to either deduplicate multiple attempts to download/delete the same thing or to mutually exclude conflicting operations
(delete while download or vice versa). The implementations of this class serve the specific use cases, like downloading the contens of a collection, or metadata of an issue, or a multistep process of downloading metadata first and then its content.

The `downloadToCache` function is now a uniform way to wait for the entirety of an collection download without needing to think about duplicated downloads. It also suspends until the process is done, and throws a single exception if failing "CacheOperationFailedException".
That should enable us to implement unhappy paths in a straight forward fashion.

All non-initiating actors that are interested in the state of a cache transaction (download indicators, progressbars(in the future?)) can listen on any opearation using a Flow object from `getCacheStatusFlow`.