Remove deprecated DataService functions and substitute with a more generic MetadataDownload
To provide a migration path before my absence I actually implemented the substitution of the deprecated DataService functions.

The solution is that the Metadata download is now a most generic function that takes any kind of observable download
and resolves the matching metadata object. Some non-happy path optimizations might have found their way into this commit as well
as it was basically in my path - like neverending spinners, because of unlimited retries when downloading.

I think this is also valuable in preparing our IssueKey problem, as the WrappedDownload now supports downloading by IssuePublication, MomentPublications and FrontPagePublications