Fix login fragment / crashes on demo days
Currently our datamodel and the functions for resolving relations only works with the assumption that any (section, article, page) resource has a 1to1 relation to an Issue.

With demo issues being public and having the same resources as the regular one that assumption is not correct anymore. As a quickfix we can
catch constraint errors on deletion (they will however be delete when the _last_ issue referencing the resource was deleted) and
return the "most valuable" issue instead of the first when resolving the issue for a section/article/page