#13901 Exclude views for to display pages and titles
The excluded views in `activity_pdf_drawer_layout` are just used for
the taz drawer. The views are used to display the front page of an
issue and the pages with titles. Since these views should just be displayed
in the taz layout they needed to be removed from `PdfPagerActivity` since
this activity is used from both build variants `LMd` and taz.