[#15373] Fix tap to scroll on Articles
Using the `setOnTouchListener()` to set our custom `WebView` touch
listener did not work on Articles, as the
`ArticleWebViewFragment.hideKeyboardOnAllViewsExceptEditText` sets a
`TouchListener` on all of its siblings. Which is unfortunately the
`AppWebView` on the article view.

Thus we did fall back to a differently named function that has to be
called explicitly.